# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task System API",
    "description": "API for managing workspaces and users within workspaces.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://yourdomain.com/api"
    }
  ],
  "paths": {
...
    "/testimonials-sections/{id}": {
      "post": {
        "tags": ["Testimonials Sections"],
        "summary": "Update a specific testimonials section by id",
        "operationId": "updateTestimonialsSection",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the testimonials section to update",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "hook": "Updated Testimonial Hook",
                "title": "Updated Testimonial Title",
                "description": "Updated testimonial description"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated testimonials section",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Testimonials section updated successfully",
                  "data": {
                    "id": 1,
                    "hook": "Updated Testimonial Hook",
                    "title": "Updated Testimonial Title",
                    "description": "Updated testimonial description",
                    "created_at": "2021-01-01T00:00:00Z",
                    "updated_at": "2021-01-01T00:00:00Z"
                  },
                  "meta": {}
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Validation error",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Validation failed",
                  "data": [],
                  "meta": {
                    "error": "Validation error message"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Unauthorized",
                  "data": [],
                  "meta": {
                    "error": "Invalid authentication token"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to update testimonials section",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Failed to update testimonials section",
                  "data": [],
                  "meta": {
                    "error": "Error message from server"
                  }
                }
              }
            }
          }
        }
      }
    },  "/testimonials": {
...
  },
...
  "security": [
    {
      "BearerAuth": []
    }
  ]
}